fix(web): stabilize today activity chart CI
