+++ docs/specs/r99mz-dashboard-today-activity-overview/SPEC.md
@@ -117,6 +117,7 @@
 - 2026-04-08: 创建 follow-up spec，冻结“今日并入活动总览 + 四段切换 + 今日分钟柱状 / 累计面积图 + merge-ready 收口”的范围与验收标准。
 - 2026-04-08: 已完成 Dashboard 页面重排、今日 KPI 嵌入、分钟级图表组件、范围持久化，以及相关 Vitest / Storybook 入口补齐。
 - 2026-04-08: 完成全量前端验证与 Storybook 视觉证据归档，并修正今日 `次数` 图中失败柱错误堆叠到正半轴的问题，确保失败柱始终以 0 轴为基线向下绘制。
+- 2026-04-08: 为 PR 收口修复跨平台午夜时间格式差异，强制分钟轴午夜显示为 `00:00`，并将今日图表数据构建逻辑拆出组件文件以满足 `react-refresh/only-export-components` lint 约束。
 
 ## Visual Evidence
